requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Deep Learning
torch>=2.0.0
torchvision
torchaudio

# Chemistry and Materials
ase>=3.22.0
pymatgen>=2022.0.0
dscribe>=2.0.0

# MLIP frameworks
fairchem-core>=1.10.0
# mace-torch>=0.3.0  # Uncomment when needed

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Experiment tracking
wandb>=0.12.0
tensorboard>=2.8.0

# Data handling
h5py>=3.6.0
openpyxl>=3.0.0
xlsxwriter>=3.0.0

# Utilities
tqdm>=4.62.0
pyyaml>=6.0
python-dotenv>=0.19.0
joblib>=1.1.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Code quality
black>=22.0.0
flake8>=4.0.0
mypy>=0.950
isort>=5.10.0

# Documentation
sphinx>=4.5.0
sphinx-rtd-theme>=1.0.0
nbsphinx>=0.8.0